+++ docs/changelog/CHANGELOG.md
@@ -1,4 +1,13 @@
-# 变更日志
-## 0.0.1
+## <small>0.0.2</small>
 
-- 初始化项目
+* 🏠style(settings): 添加git cz提交格式规范 ([a532358](https://github.com/marioliu2001/resources-vitepress/commit/a532358))
+* 🐛fix: 解决md中为找到img图片导致打包失败 ([f5a4a52](https://github.com/marioliu2001/resources-vitepress/commit/f5a4a52))
+* 🐛fix: 修复暗黑主题下title的颜色\&优化\&backTop功能 ([8a24bc6](https://github.com/marioliu2001/resources-vitepress/commit/8a24bc6))
+* feat:更新项目配置 ([7656e93](https://github.com/marioliu2001/resources-vitepress/commit/7656e93))
+* first commit ([7724a97](https://github.com/marioliu2001/resources-vitepress/commit/7724a97))
+* update:更新最新网址 ([1b759d5](https://github.com/marioliu2001/resources-vitepress/commit/1b759d5))
+* update:更新copyright和2048官网 ([9bbbc39](https://github.com/marioliu2001/resources-vitepress/commit/9bbbc39))
+* update:增加几个新网站&评分等级 ([7a6e7b5](https://github.com/marioliu2001/resources-vitepress/commit/7a6e7b5))
+
+
+